requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.3
python-dotenv==1.0.1
chromadb==0.5.23
sentence-transformers==3.3.1
python-multipart==0.0.20
